apps/negentropy-cognizes/src/cognizes/examples/e2e_travel_agent/pyproject.toml
[project]
name = "travel-agent-demo"
version = "0.1.0"
description = "Travel Agent E2E Demo for Open Agent Engine Validation"
requires-python = ">=3.13"

dependencies = [
    # ADK Core
    "google-adk>=1.22.0",               # Google Agent 开发框架 (LlmAgent, SessionService)
    "google-genai>=0.3.0",
    # Open Agent Engine (Our Implementation)
    # cognizes 包通过 uv workspace 自动引入
    "agentic-ai-cognizes",                # Workspace 依赖 (包含 cognizes 包)
    "asyncpg>=0.31.0",                  # PostgreSQL 异步驱动 (高性能, 原生 async/await)
    "psycopg[binary]>=3.2.0",           # PostgreSQL 同步驱动 (TracingManager 需要)
    "pydantic>=2.12.5",                 # 数据验证和序列化 (BaseModel, 类型提示)
    # Frontend
    "streamlit>=1.52.2",
    # Observability
    "opentelemetry-api>=1.37.0",        # 全链路追踪 API
    "opentelemetry-sdk>=1.37.0",        # 全链路追踪 SDK
    "opentelemetry-exporter-otlp>=1.37.0", # OTLP 协议导出器 (兼容 OTEL Collector)
    # Testing
    "pytest>=9.0.2",                    # Python 标准测试框架
    "pytest-asyncio>=1.3.0",            # pytest 异步测试支持 (async def test_xxx)
    "pytest-dotenv>=0.5.2",
    "microsandbox>=0.1.8",
]

[project.optional-dependencies]
dev = [
    "black>=24.4.0",
    "ruff>=0.4.0",
]

# ------------------------------------------------------------------------------
# [tool.uv.sources] - uv 依赖源配置
# 指定 workspace 依赖来源
# ------------------------------------------------------------------------------
[tool.uv.sources]
agentic-ai-cognizes = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ------------------------------------------------------------------------------
# [tool.hatch.build.targets.wheel] - Hatch 构建目标配置
# 控制 wheel 包的内容和结构
# ------------------------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["src"]             # 要打包的源码目录 (src layout 布局)
